+++ src/main/java/com/phcworld/userservice/controller/port/SuccessResponseDto.java
@@ -1,4 +1,4 @@
-package com.phcworld.userservice.dto;
+package com.phcworld.userservice.controller.port;
 
 import lombok.Builder;
