+++ src/main/java/com/phcworld/userservice/controller/port/UserService.java
@@ -1,0 +1,4 @@
+package com.phcworld.userservice.controller.port;
+
+public interface UserService {
+}
