+++ src/main/java/com/phcworld/userservice/infrastructure/UserJpaRepository.java
@@ -1,7 +1,6 @@
-package com.phcworld.userservice.repository;
+package com.phcworld.userservice.infrastructure;
 
 
-import com.phcworld.userservice.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -9,10 +8,10 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-	Optional<User> findByEmail(String email);
-	Optional<User> findByUserId(String userId);
+public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
+	Optional<UserEntity> findByEmail(String email);
+	Optional<UserEntity> findByUserId(String userId);
 
 	@Query(nativeQuery = true, value = "SELECT * FROM USERS AS U WHERE U.USER_ID IN (:userIds)")
-	List<User> findByUserId(@Param("userIds") List<String> userIds);
+	List<UserEntity> findByUserId(@Param("userIds") List<String> userIds);
 }
