+++ src/main/java/com/phcworld/userservice/jwt/service/CustomUserDetailsService.java
@@ -3,7 +3,8 @@ package com.phcworld.userservice.jwt.service;
 
 import com.phcworld.userservice.exception.model.DeletedEntityException;
 import com.phcworld.userservice.exception.model.NotFoundException;
-import com.phcworld.userservice.repository.UserRepository;
+import com.phcworld.userservice.infrastructure.UserEntity;
+import com.phcworld.userservice.infrastructure.UserJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -19,7 +20,7 @@ import java.util.Collections;
 @RequiredArgsConstructor
 public class CustomUserDetailsService implements UserDetailsService {
 
-    private final UserRepository userRepository;
+    private final UserJpaRepository userRepository;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
@@ -29,7 +30,7 @@ public class CustomUserDetailsService implements UserDetailsService {
     }
 
     // DB 에 User 값이 존재한다면 UserDetails 객체로 만들어서 리턴
-    private UserDetails createUserDetails(com.phcworld.userservice.domain.User user) {
+    private UserDetails createUserDetails(UserEntity user) {
         if(user.getIsDeleted()){
             throw new DeletedEntityException();
         }
