+++ src/main/java/com/phcworld/userservice/messagequeue/port/KafkaUserDto.java
@@ -1,4 +1,4 @@
-package com.phcworld.userservice.dto;
+package com.phcworld.userservice.messagequeue.port;
 
 import lombok.Builder;
 import lombok.Data;
