+++ src/main/java/com/phcworld/userservice/service/UserService.java
@@ -1,27 +1,20 @@
 package com.phcworld.userservice.service;
 
 import com.phcworld.userservice.domain.Authority;
-import com.phcworld.userservice.domain.User;
-import com.phcworld.userservice.dto.LoginUserRequestDto;
-import com.phcworld.userservice.dto.SuccessResponseDto;
-import com.phcworld.userservice.dto.UserRequestDto;
-import com.phcworld.userservice.dto.UserResponseDto;
+import com.phcworld.userservice.infrastructure.UserEntity;
+import com.phcworld.userservice.controller.port.SuccessResponseDto;
+import com.phcworld.userservice.domain.port.UserRequestDto;
+import com.phcworld.userservice.controller.port.UserResponseDto;
 import com.phcworld.userservice.exception.model.DuplicationException;
 import com.phcworld.userservice.exception.model.NotFoundException;
 import com.phcworld.userservice.exception.model.UnauthorizedException;
-import com.phcworld.userservice.jwt.TokenProvider;
-import com.phcworld.userservice.jwt.dto.TokenDto;
-import com.phcworld.userservice.jwt.service.CustomUserDetailsService;
 import com.phcworld.userservice.messagequeue.UserProducer;
-import com.phcworld.userservice.repository.UserRepository;
+import com.phcworld.userservice.infrastructure.UserJpaRepository;
 import com.phcworld.userservice.security.utils.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
@@ -34,18 +27,18 @@ import java.util.stream.Collectors;
 @Slf4j
 public class UserService {
 
-	private final UserRepository userRepository;
+	private final UserJpaRepository userRepository;
 	private final PasswordEncoder passwordEncoder;
 //	private final UploadFileService uploadFileService;
 	private final UserProducer userProducer;
 
-	public User registerUser(UserRequestDto requestUser) {
+	public UserEntity registerUser(UserRequestDto requestUser) {
 		userRepository.findByEmail(requestUser.email())
 				.ifPresent(user -> {
 					throw new DuplicationException();
 				});
 
-		User user = User.builder()
+		UserEntity user = UserEntity.builder()
 				.email(requestUser.email())
 				.name(requestUser.name())
 				.password(passwordEncoder.encode(requestUser.password()))
@@ -62,7 +55,7 @@ public class UserService {
 	}
 
 	public UserResponseDto getUserInfo(String userId){
-		User user = userRepository.findByUserId(userId)
+		UserEntity user = userRepository.findByUserId(userId)
 				.orElseThrow(NotFoundException::new);
 		return UserResponseDto.of(user);
 	}
@@ -72,7 +65,7 @@ public class UserService {
 		if(!userId.equals(requestDto.userId())){
 			throw new UnauthorizedException();
 		}
-		User user = userRepository.findByUserId(requestDto.userId())
+		UserEntity user = userRepository.findByUserId(requestDto.userId())
 				.orElseThrow(NotFoundException::new);
 		String profileImg = user.getProfileImage();
 		if(requestDto.imageName() != null){
@@ -96,7 +89,7 @@ public class UserService {
 		if(!securityUserId.equals(userId) && authorities != Authority.ROLE_ADMIN){
 			throw new UnauthorizedException();
 		}
-		User user = userRepository.findByUserId(userId)
+		UserEntity user = userRepository.findByUserId(userId)
 				.orElseThrow(NotFoundException::new);
 		user.delete();
 		userProducer.send("users", user);
@@ -111,7 +104,7 @@ public class UserService {
 		return userRepository.findByUserId(userIds)
 				.stream()
 				.collect(Collectors.toMap(
-						User::getUserId,
+						UserEntity::getUserId,
 					UserResponseDto::of)
 				);
 	}
