+++ src/main/java/com/phcworld/userservice/service/port/UserRepository.java
@@ -1,0 +1,4 @@
+package com.phcworld.userservice.service.port;
+
+public interface UserRepository {
+}
