+++ src/main/resources/import.sql
@@ -1,3 +1,3 @@
 insert into users(EMAIL, PASSWORD, NAME, AUTHORITY, CREATE_DATE, UPDATE_DATE, PROFILE_IMAGE, IS_DELETED, USER_ID) values('test@test.test', '$2a$10$aWqY0MzLKnt.6bvFk4zhPu.HZDabDQttLC2uAupM1yq1p6cTSTjSi', '테스트', 'ROLE_ADMIN', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(), 'blank-profile-picture.png', false, 'a2240b59-47f6-4ad4-ba07-f7c495909f40');
--- insert into users(EMAIL, PASSWORD, NAME, AUTHORITY, CREATE_DATE, PROFILE_IMAGE, IS_DELETED, USER_ID) values('test2@test.test', 'test2', '테스트2', 'ROLE_USER', CURRENT_TIMESTAMP(), 'blank-profile-picture.png', false, '3465335b-5457-4219-a0b2-d0c8b79d16ac');
--- insert into users(EMAIL, PASSWORD, NAME, AUTHORITY, CREATE_DATE, PROFILE_IMAGE, IS_DELETED, USER_ID) values('test3@test.test', 'test3', '테스트3', 'ROLE_USER', CURRENT_TIMESTAMP(), 'blank-profile-picture.png', true, '4ba22d11-66a2-483a-a255-12bcb91d8e49');
+insert into users(EMAIL, PASSWORD, NAME, AUTHORITY, CREATE_DATE, UPDATE_DATE, PROFILE_IMAGE, IS_DELETED, USER_ID) values('test2@test.test', 'test2', '테스트2', 'ROLE_USER', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(), 'blank-profile-picture.png', false, '3465335b-5457-4219-a0b2-d0c8b79d16ac');
+insert into users(EMAIL, PASSWORD, NAME, AUTHORITY, CREATE_DATE, UPDATE_DATE, PROFILE_IMAGE, IS_DELETED, USER_ID) values('test3@test.test', 'test3', '테스트3', 'ROLE_USER', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(), 'blank-profile-picture.png', true, '4ba22d11-66a2-483a-a255-12bcb91d8e49');
